Keep only volume module.
